refactor(backend): extract behavior crud module
Move the six behavior functions out of crud/__init__.py into
crud/behavior.py, following the split already done for tag, task, source
and others. _BEHAVIOR_RELATED_FIELDS moves with them -- nothing else in
__init__.py referenced it. Consumers import the module directly; nothing
is re-exported.

Nothing was dropped: all six functions still have callers.

test_test_generation_pipeline.py patched the module-level `crud` name in
test_generation_pipeline to stub get_behaviors, so those patch targets now
point at `behavior_crud`.
