+++ apps/backend/src/rhesis/backend/app/crud/__init__.py
@@ -321,124 +321,6 @@ def delete_category(
     )
 
 
-# Behavior CRUD
-# read_behavior/read_behaviors return BehaviorWithMetricsSchema (tags, user, metrics with
-# metric_type/backend_type/tags); status/organization/project are unused, excluded. metrics.tags
-# is included explicitly since selectinload's default cascade skips many-to-many relations
-# (would otherwise lazy-load tags per nested metric).
-_BEHAVIOR_RELATED_FIELDS = (
-    include(models.Behavior.user),
-    include(models.Behavior._tags_relationship, models.TaggedItem.tag),
-    include(models.Behavior.metrics),
-    include(models.Behavior.metrics, models.Metric.metric_type),
-    include(models.Behavior.metrics, models.Metric.backend_type),
-    include(models.Behavior.metrics, models.Metric._tags_relationship, models.TaggedItem.tag),
-)
-
-
-def get_behavior(
-    db: Session, behavior_id: uuid.UUID, organization_id: str = None, user_id: str = None
-) -> Optional[models.Behavior]:
-    """Get behavior with relationships eagerly loaded."""
-    from rhesis.backend.app.utils.crud_utils import _check_and_raise_if_deleted
-    from rhesis.backend.app.utils.query_utils import QueryBuilder
-
-    item = (
-        QueryBuilder(db, models.Behavior)
-        .with_deleted()
-        .with_related(*_BEHAVIOR_RELATED_FIELDS)
-        .with_organization_filter(organization_id)
-        .with_visibility_filter(user_id)
-        .filter_by_id(behavior_id)
-    )
-    return _check_and_raise_if_deleted(item, models.Behavior, behavior_id, False)
-
-
-def get_behaviors(
-    db: Session,
-    skip: int = 0,
-    limit: int = 20,
-    sort_by: str = "created_at",
-    sort_order: str = "desc",
-    filter: str | None = None,
-    organization_id: str = None,
-    user_id: str = None,
-) -> List[models.Behavior]:
-    """Get behaviors."""
-    return get_items(
-        db, models.Behavior, skip, limit, sort_by, sort_order, filter, organization_id, user_id
-    )
-
-
-def get_behaviors_detail(
-    db: Session,
-    skip: int = 0,
-    limit: int = 20,
-    sort_by: str = "created_at",
-    sort_order: str = "desc",
-    filter: str | None = None,
-    organization_id: str = None,
-    user_id: str = None,
-) -> List[models.Behavior]:
-    """Get behaviors with related objects for BehaviorWithMetricsSchema, including metrics.
-
-    Runs as two queries, mirroring get_metrics/crud_utils.get_items_detail: a joinless
-    query picks the page's IDs (filter + sort + LIMIT/OFFSET), then a second query
-    eager-loads _BEHAVIOR_RELATED_FIELDS scoped to just those IDs. Without this split,
-    Postgres would have to build every join for every matching row across the org before
-    it can sort and cut down to `limit`.
-    """
-    ordered_ids = (
-        QueryBuilder(db, models.Behavior)
-        .with_organization_filter(organization_id)
-        .with_visibility_filter(user_id)
-        .with_odata_filter(filter)
-        .with_sorting(sort_by, sort_order)
-        .with_pagination(skip, limit)
-        .ids()
-    )
-    if not ordered_ids:
-        return []
-
-    items = (
-        QueryBuilder(db, models.Behavior)
-        .with_related(*_BEHAVIOR_RELATED_FIELDS)
-        .with_organization_filter(organization_id)
-        .with_visibility_filter(user_id)
-        .query.filter(models.Behavior.id.in_(ordered_ids))
-        .all()
-    )
-
-    # WHERE id IN (...) does not preserve order -- re-apply the phase-1 sort.
-    items_by_id = {item.id: item for item in items}
-    return [items_by_id[item_id] for item_id in ordered_ids if item_id in items_by_id]
-
-
-def create_behavior(
-    db: Session, behavior: schemas.BehaviorCreate, organization_id: str = None, user_id: str = None
-) -> models.Behavior:
-    """Create behavior."""
-    return create_item(db, models.Behavior, behavior, organization_id, user_id)
-
-
-def update_behavior(
-    db: Session,
-    behavior_id: uuid.UUID,
-    behavior: schemas.BehaviorUpdate,
-    organization_id: str = None,
-    user_id: str = None,
-) -> Optional[models.Behavior]:
-    """Update behavior."""
-    return update_item(db, models.Behavior, behavior_id, behavior, organization_id, user_id)
-
-
-def delete_behavior(
-    db: Session, behavior_id: uuid.UUID, organization_id: str = None, user_id: str = None
-) -> Optional[models.Behavior]:
-    """Delete behavior."""
-    return delete_item(db, models.Behavior, behavior_id, organization_id, user_id)
-
-
 # TestSet CRUD
 def get_test_set(
     db: Session, test_set_id: uuid.UUID, organization_id: str = None, user_id: str = None
